+++ src/main/java/org/appjam/bongbaek/domain/member/controller/MemberController.java
@@ -4,12 +4,12 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.appjam.bongbaek.domain.member.dto.LoginRequest;
-import org.appjam.bongbaek.domain.member.dto.LoginResponse;
-import org.appjam.bongbaek.domain.member.dto.ReissueRequest;
-import org.appjam.bongbaek.domain.member.dto.SignUpRequest;
-import org.appjam.bongbaek.domain.member.dto.UpdateMemberRequest;
-import org.appjam.bongbaek.domain.member.dto.WithdrawRequest;
+import org.appjam.bongbaek.domain.member.dto.request.LoginRequest;
+import org.appjam.bongbaek.domain.member.dto.request.LoginResponse;
+import org.appjam.bongbaek.domain.member.dto.request.ReissueRequest;
+import org.appjam.bongbaek.domain.member.dto.request.SignUpRequest;
+import org.appjam.bongbaek.domain.member.dto.request.UpdateMemberRequest;
+import org.appjam.bongbaek.domain.member.dto.request.WithdrawRequest;
 import org.appjam.bongbaek.domain.member.service.MemberService;
 import org.appjam.bongbaek.global.api.ApiResponse;
 import org.appjam.bongbaek.global.api.ApiResponse.EmptyBody;
