+++ src/main/java/org/appjam/bongbaek/domain/member/dto/request/UpdateMemberRequest.java
@@ -1,4 +1,4 @@
-package org.appjam.bongbaek.domain.member.dto;
+package org.appjam.bongbaek.domain.member.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
